backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
cryptography==44.0.0
google-genai>=1.0.0
playwright==1.49.1
pydantic==2.10.4
pydantic-settings==2.7.1
websockets==14.1
